fix(mantine): override get input props types
