+++ packages/mantine/src/index.ts
@@ -1,6 +1,7 @@
 import {MantineColorsTuple} from '@mantine/core';
 
 import {noop} from '@mantine/core';
+import {LooseKeys} from '@mantine/form/lib/types';
 import {PlasmaColors} from './theme/PlasmaColors';
 
 export * from '@mantine/carousel';
@@ -42,4 +43,30 @@ declare module '@mantine/core' {
         // eslint-disable-next-line @typescript-eslint/ban-types
         colors: Record<keyof typeof PlasmaColors | (string & {}), MantineColorsTuple>;
     }
+}
+
+declare module '@mantine/form' {
+    export type GetInputPropsType = 'input' | 'checkbox' | 'collection';
+    export interface GetInputPropsOptions {
+        type?: GetInputPropsType;
+        withError?: boolean;
+        withFocus?: boolean;
+        [key: string]: any;
+    }
+    export interface GetInputPropsReturnType {
+        onChange: any;
+        value?: any;
+        checked?: any;
+        error?: any;
+        onFocus?: any;
+        onBlur?: any;
+        onReorderItem?: any;
+        onRemoveItem?: any;
+        onInsertItem?: any;
+    }
+
+    export type GetInputProps<Values> = <Field extends LooseKeys<Values>>(
+        path: Field,
+        options?: GetInputPropsOptions,
+    ) => GetInputPropsReturnType;
 }
